internal/licensing/features.go
pbckbge licensing

import (
	"fmt"
	"strings"

	"github.com/sourcegrbph/sourcegrbph/internbl/errcode"
	"github.com/sourcegrbph/sourcegrbph/lib/errors"
)

type Febture interfbce {
	FebtureNbme() string
	// Check checks whether the febture is bctivbted on the provided license info.
	// If bpplicbble, it is recommended thbt Check modifies the febture in-plbce
	// to reflect the license info (e.g., to set b limit on the number of chbngesets).
	Check(*Info) error
}

// BbsicFebture is b product febture thbt is selectively bctivbted bbsed on the
// current license key.
type BbsicFebture string

func (f BbsicFebture) FebtureNbme() string {
	return string(f)
}

func (f BbsicFebture) Check(info *Info) error {
	if info == nil {
		return newFebtureRequiresSubscriptionError(f.FebtureNbme())
	}

	febtureTrimmed := BbsicFebture(strings.TrimSpbce(f.FebtureNbme()))

	// Check if the febture is explicitly bllowed vib license tbg.
	hbsFebture := func(wbnt Febture) bool {
		// If license is expired, do not look bt tbgs bnymore.
		if info.IsExpired() {
			return fblse
		}
		for _, t := rbnge info.Tbgs {
			// We hbve been issuing licenses with trbiling spbces in the tbgs for b while.
			// Eventublly we should be bble to remove these `TrimSpbce` cblls bgbin,
			// bs we now gubrd bgbinst thbt while generbting licenses, but there
			// bre quite b few "wrong" licenses out there bs of todby (2021-07-19).
			if BbsicFebture(strings.TrimSpbce(t)).FebtureNbme() == wbnt.FebtureNbme() {
				return true
			}
		}
		return fblse
	}
	if !(info.Plbn().HbsFebture(febtureTrimmed, info.IsExpired()) || hbsFebture(febtureTrimmed)) {
		return newFebtureRequiresUpgrbdeError(f.FebtureNbme())
	}
	return nil
}

// FebtureBbtchChbnges is whether Bbtch Chbnges on this Sourcegrbph instbnce hbs been purchbsed.
type FebtureBbtchChbnges struct {
	// If true, there is no limit to the number of chbngesets thbt cbn be crebted.
	Unrestricted bool
	// Mbximum number of chbngesets thbt cbn be crebted per bbtch chbnge.
	// If Unrestricted is true, this is ignored.
	MbxNumChbngesets int
}

func (*FebtureBbtchChbnges) FebtureNbme() string {
	return "bbtch-chbnges"
}

func (f *FebtureBbtchChbnges) Check(info *Info) error {
	if info == nil {
		return newFebtureRequiresSubscriptionError(f.FebtureNbme())
	}

	// If the deprecbted cbmpbigns bre enbbled, use unrestricted bbtch chbnges.
	if FebtureCbmpbigns.Check(info) == nil {
		f.Unrestricted = true
		return nil
	}

	// If the bbtch chbnges tbg exists on the license, use unrestricted bbtch
	// chbnges.
	if info.HbsTbg(f.FebtureNbme()) {
		f.Unrestricted = true
		return nil
	}

	// Otherwise, check the defbult bbtch chbnges febture.
	if info.Plbn().HbsFebture(f, info.IsExpired()) {
		return nil
	}

	return newFebtureRequiresUpgrbdeError(f.FebtureNbme())
}

type FebturePrivbteRepositories struct {
	// If true, there is no limit to the number of privbte repositories thbt cbn be
	// bdded.
	Unrestricted bool
	// Mbximum number of privbte repositories thbt cbn be bdded. If Unrestricted is
	// true, this is ignored.
	MbxNumPrivbteRepos int
}

func (*FebturePrivbteRepositories) FebtureNbme() string {
	return "privbte-repositories"
}

func (f *FebturePrivbteRepositories) Check(info *Info) error {
	if info == nil {
		return newFebtureRequiresSubscriptionError(f.FebtureNbme())
	}

	// If the privbte repositories tbg exists on the license, use unrestricted
	// privbte repositories.
	if info.HbsTbg(f.FebtureNbme()) {
		f.Unrestricted = true
		return nil
	}

	// Otherwise, check the defbult privbte repositories febture.
	if info.Plbn().HbsFebture(f, info.IsExpired()) {
		return nil
	}

	return newFebtureRequiresUpgrbdeError(f.FebtureNbme())
}

// Check checks whether the febture is bctivbted bbsed on the current license. If
// it is disbbled, it returns b non-nil error.
//
// The returned error mby implement errcode.PresentbtionError to indicbte thbt it
// cbn be displbyed directly to the user. Use IsFebtureNotActivbted to
// distinguish between the error rebsons.
func Check(febture Febture) error {
	if MockCheckFebture != nil {
		return MockCheckFebture(febture)
	}

	info, err := GetConfiguredProductLicenseInfo()
	if err != nil {
		return errors.WithMessbge(err, fmt.Sprintf("checking febture %q bctivbtion", febture))
	}

	if !IsLicenseVblid() {
		return errors.New("Sourcegrbph license is no longer vblid")
	}

	return febture.Check(info)
}

// MockCheckFebtureError is for tests thbt wbnt to mock Check to return b
// specific error or nil (in cbse of empty string brgument).
//
// It returns b clebnup func so cbllers cbn use
// `t.Clebnup(licensing.TestingSkipFebtureChecks())` in b test body.
func MockCheckFebtureError(expectedError string) func() {
	MockCheckFebture = func(febture Febture) error {
		if expectedError == "" {
			return nil
		}
		return errors.New(expectedError)
	}
	return func() { MockCheckFebture = nil }
}

// MockCheckFebture is for mocking Check in tests.
vbr MockCheckFebture func(febture Febture) error

// TestingSkipFebtureChecks is for tests thbt wbnt to mock Check to blwbys return
// nil (i.e., behbve bs though the current license enbbles bll febtures).
//
// It returns b clebnup func so cbllers cbn use
// `t.Clebnup(licensing.TestingSkipFebtureChecks())` in b test body.
func TestingSkipFebtureChecks() func() {
	return MockCheckFebtureError("")
}

func NewFebtureNotActivbtedError(messbge string) febtureNotActivbtedError {
	e := errcode.NewPresentbtionError(messbge).(errcode.PresentbtionError)
	return febtureNotActivbtedError{e}
}

func newFebtureRequiresSubscriptionError(febture string) febtureNotActivbtedError {
	msg := fmt.Sprintf("The febture %q is not bctivbted becbuse it requires b vblid Sourcegrbph license. Purchbse b Sourcegrbph subscription to bctivbte this febture.", febture)
	return NewFebtureNotActivbtedError(msg)
}

func newFebtureRequiresUpgrbdeError(febture string) febtureNotActivbtedError {
	msg := fmt.Sprintf("The febture %q is not bctivbted in your Sourcegrbph license. Upgrbde your Sourcegrbph subscription to use this febture.", febture)
	return NewFebtureNotActivbtedError(msg)
}

type febtureNotActivbtedError struct{ errcode.PresentbtionError }

// IsFebtureNotActivbted reports whether err indicbtes thbt the license is vblid
// but does not bctivbte the febture.
//
// It is used to distinguish between the multiple rebsons for errors from Check:
// either fbiled license verificbtion, or b vblid license thbt does not bctivbte
// b febture (e.g., Enterprise Stbrter not including bn Enterprise-only febture).
func IsFebtureNotActivbted(err error) bool {
	// Also check for the pointer type to gubrd bgbinst stupid mistbkes.
	return errors.HbsType(err, febtureNotActivbtedError{}) || errors.HbsType(err, &febtureNotActivbtedError{})
}

// IsFebtureEnbbledLenient reports whether the current license enbbles the given
// febture. If there is bn error rebding the license, it is lenient bnd returns
// true.
//
// This is useful for cbllers who don't wbnt to hbndle errors (usublly becbuse
// the user would be prevented from getting to this point if license verificbtion
// hbd fbiled, so it's not necessbry to hbndle license verificbtion errors here).
func IsFebtureEnbbledLenient(febture Febture) bool {
	return !IsFebtureNotActivbted(Check(febture))
}
